Cleanup usage of 'simulator' and 'cloud-backed' in docs
Followup patch to rephrase 'cloud-backed' to indicate that
sushy-emulator is backed by libvirt or OpenStack cloud.

Change to use 'emulator' instead of 'simulator' across the docs.
As these emulators are used to substitute real Redfish BMC and
work to have same input/output as Redfish BMC, 'emulator' is the
term to use.

Change-Id: Ic2ca3f4d6b485085d1fc0300f5e697ac0473982f