Fix nova unit tests
Somehow some of the mocks in the mocking harness
escaped to the wild and lived their own life. They
got away with that just because of the accidental loss
of `assert_` piece of the function name that is supposed
to nail those loose mocks down.

Specifically, in nova unit tests:

* Moved `openstack.connect` mocking to the class level
* Fixed improper `power_state` value
* Set up proper chain of mocks to address `.compute`
* Fixed mocks assertions

This is a follow-up of change ID:

    I82032fd13a74d9aa77616dac3802d506b4a9c1cf

Change-Id: If8a479ae793c1616ef8b30e76bd8cdfdfc08dbe0