Make external dependencies optional
This change makes driver dependencies optional thus allowing
sushy-tools to install and run test suite even if some of
these not-always-required dependencies would not otherwise
be met failing the entire installation run.

If the patch is merged, we would need to update sushy-tools
users (mainly, CI jobs) to install the drivers dependencies
of their choice or just all of them.

Story: 2004342
Task: 27929
Change-Id: Ibd05a19d66f372eaa8fc916af98722f7fe701f3c