Add configuration file to sushy-emulator
Prior to this commit, `sushy-emulator` could only be configured
through command-line options and a couple of environment variables.
This proved to be inefficient with the growth of functionality
and configuration options to support it.

This commit introduces configuration file support based entirely
on Flask's configuration file feature. Presently supported
`sushy-emulator` command-line options have been reflected in the
configuration file.

Flask app's options could also be set in the same configuration file
as described in Flask documentation.

Story: 2004305
Task: 27866
Change-Id: Ie6ce173cd8d72874cd6da141b7164f20ebaf0f3f