Add configurable libvirt firmware
When libvirt boots up a domain in UEFI mode, the system needs
a firmware loader to proceed. Apparently, the location of
the firmware blobs are not standardized across Linux distributions.

With this change, whenever dynamic emulator boots a domain in
UEFI mode, emulator configures libvirt domain to use specific
firmware image. The map of firmware loader locations can
be configured via the configuration file.

Change-Id: I15ff24d8624272c6792cc53dc9a34b69ece480b0