Fix lower-constraint deps handling
When the lower-constraints tox target was added, it was assumed the
install_command was just running the install and that the dependencies
and constraints were being set using "deps = ".

This fixed the install_command and deps to follow the expected pattern
so the lower-constraints job actual does install the lower constraints.

Raising libvirt-python minimum version as the existing version would
fail.

Change-Id: Icbbae373525dce3f13fcb38373d4c66160cf05d1
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>