Add Simple Storage resource support
Simple Storage resource is a storage model used to represent the
properties of a storage controller and its directly attached
devices. As of this patch, the Simple Storage emulation is only
supported for the libvirt driver for storage devices that are
configured as a Volume via libvirt and attached to a VM/domain.

Story: #2005948
Task: #34313
Change-Id: If10bd04d517600c3a69bb43fe0ea50f2efd62422