Fix libvirt set-boot-image implementation
When libvirt driver performs set-boot-image operation, it actually
adds new block device to libvirt domain. This new 'disk' device
should be mapped on a virtual bus and possess unique bus identifiers.

Prior to this fix, these identifiers were hardcoded in the emulator,
however it turned out that static values may sometimes collide with
already existing devices on the same bus.

It seems that libvirt is capable of automatically choosing free bus
identifiers if they are not given explicitly. So the essence of this
fix is just to let libvirt pick hopefully proper mapping.

Change-Id: I19fdae4417391a2fb753ccdd0a87fe88b4184161