Fix libvirt CD image insertion
Sometimes libvirt refuses to take our freshly burnt ISO image
and put it into /dev/cdrom. The error is:

  Found duplicate drive address for disk with target name 'vda'
  controller='0' bus='0' target='0' unit='1'

The problem seems to be that despite libvirt docs spare the user
from assigning `<address>` to device, sometimes libvirt fails to
find a free spot on the bus.

This patch attempts to enumerate already existing devices on the
bus and pick a free `unit`. Then we will craft a seemingly proper
address and add it to `disk` libvirt configuration.

Change-Id: I5975e0cd016ede7a5685eaca34c19af295f7f172