Stop configuring install_command in tox.
Currently, we are overriding 'install_command' to use 'pip'. This is
considered poor behavior and 'python -m pip' should be used instead:

https://snarky.ca/why-you-should-use-python-m-pip/

It turns out that this is the the default value provided by tox:

https://tox.readthedocs.io/en/latest/config.html#conf-install_command

So we can remove the line and simply use the default value.

openstack-discuss thread
http://lists.openstack.org/pipermail/openstack-discuss/2020-April/014237.html

Bumping l-c:
-openstacksdk to 0.13.0
-Werkzeug to 0.14

Bumping t-r:
-openstacksdk to 0.13.0

Change-Id: I6e7e3792f2cede28ea9098ba546f415cec0ddf9b