Replace the static Manager driver with a new fake one
Refactors the Manager interface to provide a single get_manager call
instead of uuid/name calls that transform identities.

Add new calls to list systems and chassis per manager.

Do not use jinja for rendering JSON, it's redundant, just use Python.

Finally, provide a manager driver that ties a manager to a system
in a 1-1 relationship. This fixes the current issue that one manager
managers all systems, and thus a virtual media is always attached
to all systems.

Change-Id: I46dd794750c182572d37bab3ebfcd608ceb1bb6c