Return an integer value for Cores and Threads
Clients like Ironic Redfish inspector expect an integer value for
Cores and Threads. We do our best to return an integer value, we also
provide a default of "1" in case none is found.

Current logic expects the number of threads and cores to be present
in the domain, but this is not necessarily the case (test case
provided). We also can return a lot of information even when the
domain is not active, so we do just that.

Story: 2008743
Task: 42100

Change-Id: Id702ca2d7c5965217f348fc0931df0b0b1e57da8