Use /dev/sdx instead of /dev/sdc when mounting boot images
Currently when mounting cdrom device we are hardcoding use of /dev/sdc.
Given that in some scenarios users may want to attach additional disks
to their VMs, it means they need to be aware that using /dev/sdc is not
permitted if e.g. there is an inspector image mounted as Virtual Media.

In order to simplify the usage we are instead mounting boot devices as
/dev/sdx so that users have a freedom of creating block devices without
discontinuous names.

Change-Id: Ifd072ca6e0a88da138d9ef7140499faca1705c46