Add a fake system driver
I would like to scale-test Ironic in presence of many nodes. To make it
possible without creating many VMs, this new driver allows having fake
Redfish systems that react to the API, but do not do anything.

This change is enough to make the "ramdisk" deploy work. As the next
step, we can introduce a fake IPA implementation that allows the normal
deploy to "work" (as in: finish successfully).

Change-Id: I23f5472471dd4faf78a067d490d785b3b15db42b