Fix the CI
Lower constraints do not cap all dependencies, causing incompatibilities.
Update Jinja2 to 2.11.3 (the last pre-3.0 version) and downgrade
MarkupSafe, Werkzeug and itsdangerous for Flask.

Something must have changed in one of the dependencies, and the Location
header is no longer returning the host in the URL.

Remove subunit and testscenarios, we don't use them directly.

Change-Id: I5bad754ef7b1bec6986c2b4293a38226c5a261af