Fix per-request memoization
The original intent[1] of memoization was for @memoize.memoize()
decorated methods to be memoized for the duration of one request,
however nothing ever reset the _cache dict so memoization persisted
for the duration of the application.

This revealed itself because fetching one system with the nova driver
returned very quickly and the PowerState never changed after a
Actions/ComputerSystem.Reset POST to switch it on or off.

This change resets app._cache at the beginning of each request. This
won't affect @memoize.memoize(permanent_cache=PERMANENT_CACHE)
behaviour.

[1] https://review.opendev.org/c/openstack/sushy-tools/+/612758

Change-Id: Ic1691dc988969aeda85f44f9014e43dfa6484c68