aarch64/arm64 libvirt VMs should default to SCSI controller
The current logic breaks when sushy-tools tries to mount virtual media to an arm64/aarch64 libvirt vm.
This is because by default it tries to connect and ide controller, and the only controller option for arm is scsi.

This patch ensures that aarch64 type vms will always get a scsi controller.

Story: 2009962
Task: 44940
Change-Id: Ib7c3267f401355cf045954e71fe7297b484508be