Save previous hd boot device when setting Pxe
Assuming that sushy_tool ignores "BootSourceOverrideEnabled": "Once" for
libvirt backend, we define a new boot device permanently. And if the pxe
server does not reply after the first successful  provision, we just get a
vm which doesn't manage to boot after provision because netboot is the only
option and no reply from pxe.

This patch adds two additional lists of previously configured boot hard drives
for the vm at the end of the boot order. The first list is for the os section
If it is defined, it just adds all hard drives to the list. The second list is
for the actual drive section with the boot directive. 

Tried to save script logic as much as possible. But it might be reasonable to
pass one additional argument to the set_boot_device function to do such
altering when "BootSourceOverrideEnabled" is defined and has "Once" as a value

Added testcase to verify boot order definition

Change-Id: I0abcab9ccbffcc9b06382b0e6d335edd20db25bf