Simplify UEFI logic and change the UefiHttp flow
In trying to test the Redfish UEFI Http interface logic, it was
observed that we were not getting the values out of the emulator we
expected. So this change simplifies the logic which wasn't quite
working for us.

We also realized the code was trying to set the boot interface before
actually creating it. Doh!

It seems, also I might have had some other issues in the original patch
which we couldn't detect without fully getting everything working, so
some additional refactoring exists with this patch.

Change-Id: Ibfe5a560a25251965e8bad87012f64e89d9640be