Use a UUID instead of a named identity for httpboot
This is an alternative to https://review.opendev.org/c/openstack/sushy-tools/+/902015
The tl;dr is we can leak the uuid out in some weird cases and
while things will just keep working, some methods internally
really just want a UUID and not a name. The other patch
is the name fix.

This, explicitly uses a UUID for httpboot, in the event
it is still happilly working with a name up until that point.

Change-Id: I52121f7e6148aaa3739bc9900b4480b7b3dff8fc