Use existing nvram path in libvirt driver to support session daemon mode
When libvirtd is started as a non-root user it does not have access to
folder /var/lib/libvirt. By default, libvirtd in session daemon mode
creates a nvram file at path ~/.config/libvirt/qemu/nvram/\
[DOMAIN]_VARS[EXT] instead [0],[1].

Querying the nvram path in session daemon mode is difficult because
libvirtd does not expose it nor $HOME of the user that runs the daemon.
As a workaround, this change will let libvirtd create the nvram file
path if it has not been previously defined [2]. However, if it had been
defined earlier then the path will be changed to enforce regenerating
it from the template.

[0] https://github.com/libvirt/libvirt/blob/2b148748c223dd94f42b91a03e7e4c5ecfa8a5b2/src/qemu/qemu_conf.c#L203
[1] https://github.com/libvirt/libvirt/blob/2b148748c223dd94f42b91a03e7e4c5ecfa8a5b2/src/qemu/qemu_firmware.c#L1124
[2] https://github.com/libvirt/libvirt/blob/2b148748c223dd94f42b91a03e7e4c5ecfa8a5b2/src/qemu/qemu_firmware.c#L1065

Change-Id: Id78101f5c1f5fdf4b365d2d270dbde8877b05dbf