Extend the fake systems driver with external systems status notifier
The fake driver can now share the status of fake systems with any external component expected to run on the VMs. This update adds a function to notify an external API when the power status of a fake VM changes. The notification is sent via a PUT API method with the system object as a JSON parameter. This allows the receiver to implement an endpoint that triggers some process based on the new status.

Change-Id: I8897c4ce03fc02b3fec4152ff6d52a4e8e81013a
Signed-off-by: Mohammed Boukhalfa <mohammed.boukhalfa@est.tech>