Openstack vmedia - add glance file upload support
The openstack driver for virtual media now supports Glance file upload.

As an alternative to using image import via the ``web-download`` method,
it is now possible to configure the emulator to download the image to a
temp file which is then uploaded to Glance.

Since the vmedia drivers OpenstackDriver and StaticDriver now both use
requests get() to download the image file - the code has been
re-factored to a shared function _get_image in the vmedia BaseDriver.

Change-Id: Ied050dda997e5b6acbfb3acf11cb3348b697fb3a