Openstack vmedia - refactor to pre-defined volumes
... I hit the wall with the rebuild from image volume when instances
where UEFI. The rebuild from volume image failed with this exception
when UEFI image properties where populated.
  libvirt.libvirtError: Requested operation is not valid: cannot
                        undefine domain with nvram

This changes the approach, instead of creating and attaching the
volume the Nova instance is pre-confonfigured with two volumes.
One is device_type: `disk` and the other device_type: `cdrom` on
`scsi` bus. The cdrom volume is configured with boot_index: 0, and
the disk with boot_index: 1

The image used for the `cdrom` device initially, is a non-bootable
"blank-image", the result is that the instance fall back to the
`disk` device when booting.

After insert + set_boot_image the server is rebuilt with the inserted
image, this rebuild will only re-image the volume with `boot_index: 0`
and so following reboot the instance will boot of the `cdrom` volume.
After eject + set_boot_image the server is rebuilt using the
"blank-image", following reboot will boot of the `disk` volume.

Change-Id: I16f3fcf79f34b8288e6b8c107fd49bbd7acd4b51