Openstack - Raise SYS518 server.task_state != None
The Openstack driver's vmedia support relies on rebuilding nova
instances, this process runs in the background and may take some time.
If ironic tries to perform a power action on an instance which is
rebuilding Nova will raise an error.

This change adds a check for ``task_state`` on the nova instance, and
raises 503 with ``SYS518`` in the message if ``task_state`` is not
``None``.

The error ``SYS518`` code is also used by iDRAC to signal that it is
"busy" with some task and sushy already does retries for this error.

Change-Id: Ie81362b8c0d026706e1e45c8b821a832f847bb16